everything compiles now...need to start testing. PolygonOperations.Simplify.cs has had some methods merely commented though

+++ TessellationAndVoxelizationGeometryLibrary/Polygon Operations/PolygonBooleanOperations/PolygonRemoveIntersections.cs
@@ -78,7 +78,7 @@ namespace TVGL
                 currentEdge = null;
                 return false;
             }
-            if (int.IsEvenInteger(intersectionData.Relationship)) // must be odd. otherwise it is a noa a ValidOverlap
+            if (int.IsEvenInteger((int)intersectionData.Relationship)) // must be odd. otherwise it is a noa a ValidOverlap
             {
                 currentEdge = null;
                 return false;
